requirements.txt
pandas
numpy
scikit-learn
skforecast
xgboost
lightgbm
matplotlib
seaborn
ipykernel
openpyxl
statsmodels
tqdm
python-dotenv

supabase
